+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
 
+- concord-runner: fix `JavaDelegate` handling;
 - concord-server: fix SSH key pair upload/create endpoint;
 - concord-server: process events cleanup;
 - concord-server: fixed a potential NPE while retrieving the process